+++ backend-admin/app/Http/Controllers/RoleController.php
@@ -6,29 +6,21 @@ use App\Models\Role;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Http\Resources\RoleResource;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 
 class RoleController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
+        Gate::authorize('view', 'roles');
         return RoleResource::collection(Role::all());
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
-    {
+    {        
+        Gate::authorize('edit', 'roles');
         $role = Role::create($request->only('name'));
 
         if ($permissions = $request->input('permissions')) {
@@ -42,27 +34,16 @@ class RoleController extends Controller
         return response(new RoleResource($role), Response::HTTP_CREATED);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
+        Gate::authorize('view', 'roles');
         $role = Role::find($id);
         return new RoleResource($role);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'roles');
         $role = Role::find($id);
         $role->update($request->only('name')); 
         
@@ -74,19 +55,13 @@ class RoleController extends Controller
                     'permission_id' => $permission_id
                 ]);
             }    
-        }        
-        
+        }
         return response($role, Response::HTTP_ACCEPTED); // 202
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($id)
     {
+        Gate::authorize('edit', 'roles');
         DB::table('role_permission')->where('role_id', $id)->delete();
         Role::destroy($id);        
         return response(null, Response::HTTP_NO_CONTENT);
