+++ backend-admin/app/Http/Controllers/UserController.php
@@ -5,31 +5,36 @@ namespace App\Http\Controllers;
 use Auth;
 use App\Models\User;
 use Illuminate\Http\Request;
+use App\Http\Resources\UserResource;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\UpdateInfoRequest;
 use App\Http\Requests\UserCreateRequest;
 use App\Http\Requests\UserUpdateRequest;
 use App\Http\Requests\UpdatePasswordRequest;
-use App\Http\Resources\UserResource;
 use Symfony\Component\HttpFoundation\Response;
 
 class UserController extends Controller
 {
     public function index() {
+        Gate::authorize('view', 'users');
         //return User::with('role')->paginate(5);
-        // Use UserResource instead of 'with'
+        // Use UserResource instead of 'with'        
         $users = User::paginate(5);
         return UserResource::collection($users);
     }
 
     public function show($id) {
+        Gate::authorize('view', 'users');
         //$user = User::with('role')->find($id);
-        // Use UserResource instead of 'with'
+        // Use UserResource instead of 'with'        
         $user = User::find($id);
         return new UserResource($user);
     }
 
     public function store(UserCreateRequest $request) {
+
+        Gate::authorize('edit', 'users');
         //$user = User::create($request->all());
         $user = User::create($request->only('first_name', 'last_name', 'email', 'role_id') + [
             'password' => Hash::make('password'), // Default password given by system
@@ -38,12 +43,14 @@ class UserController extends Controller
     }
 
     public function update(UserUpdateRequest $request, $id) {
+        Gate::authorize('edit', 'users');
         $user = User::find($id);
         $user->update($request->only('first_name', 'last_name', 'email', 'role_id'));
         return response(new UserResource($user), Response::HTTP_ACCEPTED); // 202
     }
 
     public function delete($id) {
+        Gate::authorize('edit', 'users');
         User::destroy($id);
         return response(null, Response::HTTP_NO_CONTENT);
     }
